pyproject.toml
[tool.poetry]
name = "iscc-registry"
version = "0.1.0"
description = "ISCC - Decentralized Content Registry"
authors = ["Titusz <tp@py7.de>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/iscc/iscc-registry"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
Django = "^4.0"
django-environ = "^0.8.1"
loguru = "^0.6.0"
django-ninja = "^0.17.0"
iscc-core = "^0.2.4"
jcs = "^0.2.0"
django-json-widget = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
black = "^22.3"
poethepoet = "^0.13"
pytest-django = "^4.5.2"

[tool.black]
skip-string-normalization = false
line-length = 100
target-version = ['py38']

[tool.poe.tasks]
demo = { cmd = "poetry run python -m dev.demo", help = "Install app with demo user"}
load = { cmd = "poetry run python -m dev.load", help = "Load some test data"}
format = { cmd = "poetry run black .", help = "Code style formating with black" }
test = { cmd = "poetry run pytest", help = "Run tests" }
all = ["format", "test"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
